update: migrate constants to TypeScript and update import paths
