+++ src/api/index.js
@@ -1,7 +1,7 @@
 import fs from 'node:fs/promises'
 import { dirname, join } from 'node:path'
 import { fileURLToPath } from 'node:url'
-import { PATH_VERSION } from '../constants.js'
+import { PATH_VERSION } from '../constants.ts'
 import {
   logger,
   sendErrorResponse,
