+++ src/playback/processing/streamProcessor.js
@@ -6,7 +6,7 @@ import FAAD2NodeDecoder from '@ecliptia/faad2-wasm/faad2_node_decoder.js'
 import { SeekError, seekableStream } from '@ecliptia/seekable-stream'
 import { SymphoniaDecoder } from '@toddynnn/symphonia-decoder'
 import * as MP4Box from 'mp4box'
-import { normalizeFormat, SupportedFormats } from '../../constants.js'
+import { normalizeFormat, SupportedFormats } from '../../constants.ts'
 import { logger } from '../../utils.js'
 import FlvDemuxer from '../demuxers/Flv.js'
 import WebmOpusDemuxer from '../demuxers/WebmOpus.js'
