+++ src/workers/main.js
@@ -2,7 +2,7 @@ import net from 'node:net'
 import os from 'node:os'
 import { monitorEventLoopDelay } from 'node:perf_hooks'
 import v8 from 'node:v8'
-import { GatewayEvents } from '../constants.js'
+import { GatewayEvents } from '../constants.ts'
 import ConnectionManager from '../managers/connectionManager.js'
 import CredentialManager from '../managers/credentialManager.js'
 import TrackCacheManager from '../managers/trackCacheManager.js'
